Alignments refactor
