+++ .gitignore
@@ -19,3 +19,4 @@
 !scripts
 !plugins
 !tools
+!tools/lib*
